[REFACTORING] - Update token submodule name
[*] Update submodule name to be compliant with clippy
